Slightly more than basic youtube downloader
This is a more than basic YouTube down loader. It downloads a video in 2 separate pieces, the actual video sans audio, and then the audio. This is so that the best high quality versions of both can be downloaded without interference of the other when using WiFi. Then both files are stitched together.